requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
decorator==5.1.0
et-xmlfile==1.1.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
openpyxl==3.0.9
pbr==5.6.0
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
Tempita==0.5.2
urllib3==1.26.7
Werkzeug==2.0.2
